+++ src/helpers.ts
@@ -7,3 +7,20 @@ export function getHashFromURL(url: string | URL) {
 
   return null;
 }
+
+export function readFileAsArrayBuffer(file: Blob | File) {
+  return new Promise<ArrayBuffer>((resolve, reject) => {
+    const reader = new FileReader();
+    reader.onload = (e) => {
+      const result = e.target?.result;
+      if (result == undefined || typeof result !== "object") {
+        reject();
+        return;
+      }
+
+      resolve(result);
+    };
+    reader.onerror = (error) => reject(error);
+    reader.readAsArrayBuffer(file);
+  });
+}