+++ CHANGELOG.md
@@ -4,4 +4,14 @@
 
 ### Minor Changes
 
+- Include "x" tag with sha256 in upload auth
+
+### Patch Changes
+
+- Allow URL class for server
+
+## 0.7.0
+
+### Minor Changes
+
 - e2e93f5: Add `handleImageFallbacks` method